requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
TA-Lib>=0.4.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0


